requirements.txt
# ============================================================
#  DataPattern — Offer Letter Agent
#  Python dependencies
#
#  Install with:
#    pip install -r requirements.txt
# ============================================================

# ── Web framework ────────────────────────────────────────
streamlit>=1.35.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── Standard library modules used (no pip install needed) ─
# os, re, zipfile, tempfile, time, pathlib, smtplib,
# email.mime.* — all built into Python 3.10+

# ── Optional: type hints support (Python < 3.10) ─────────
# If you are on Python 3.9 or below, uncomment:
# typing_extensions>=4.0.0
